Fix cloning on branches. Show template on tasks
