+++ third-party-extensions.md
@@ -3,7 +3,6 @@ Allow 3-party extensions to be installed on the SE-2 instance.
 ToDo:
 
 - [x] Read arguments (github URL, or maybe just org/repo:branch)
-- [] Fetch repo
-  - [x] Clone repo
-  - [] Check if it's a valid repo
-  - [] Do templating stuff.
+- [x] Validate / Clone repo
+- [] Do templating on repo template
+- [] Warning on 3party templates.
